Rename: folderCardCollection의 이름을 컴포넌트의 역할에 맞춰 재지정
