+++ src/components/FolderShareCollection.css
