+++ src/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import Folder from "components/Folder";
+import FolderShare from "components/FolderShare";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import App from "components/App";
 
@@ -9,7 +9,7 @@ root.render(
 	<BrowserRouter>
 		<Routes>
 			<Route path="/" element={<App />}>
-				<Route path="share" element={<Folder />} />
+				<Route path="share" element={<FolderShare />} />
 			</Route>
 		</Routes>
 	</BrowserRouter>
